requirements.txt
PyYAML >= 5.3.1
logzero >= 1.5.0